[admin] refactor(209): upload-approval entity를 전용 폴더로 구조화
상품 관련 entity 파일들을 도메인별로 정리하여 FSD 아키텍처를 개선했습니다.

- entity/product 하위 파일들을 entity/product/upload-approval 폴더로 이동
- product.* 파일명을 upload-approval.*로 변경
- index.ts에서 export 구조 정리 및 Product 타입은 유지
