+++ apps/admin/src/entity/product/upload-approval/upload-approval.query.ts
@@ -1,8 +1,8 @@
 import { queryOptions } from '@tanstack/react-query'
 
-import { getUploadApprovals } from './product.api'
+import { getUploadApprovals } from './upload-approval.api'
 
-import type { GetUploadApprovalsRequestParams } from './product.type'
+import type { GetUploadApprovalsRequestParams } from './upload-approval.type'
 
 export const productQueries = {
   all: () => ['products'] as const,
